fix shared get_document dependency
