+++ backend/app/core/db.py
@@ -1,7 +1,3 @@
-from fastapi import Depends, HTTPException
-from sqlmodel import select
-from app.models import Document
-from sqlalchemy.exc import NoResultFound
 from sqlmodel import create_engine, Session
 import logging
 
@@ -21,17 +17,3 @@ engine = create_engine(str(settings.SQLALCHEMY_DATABASE_URI), echo=echo)
 def get_session():
     with Session(engine) as session:
         yield session
-
-
-def get_document(document_id: str, session: Session = Depends(get_session)) -> Document:
-    try:
-        document = session.exec(
-            select(Document).where(Document.document_id == document_id)
-        ).one()
-    except NoResultFound:
-        raise HTTPException(status_code=404, detail="Document not found")
-    except Exception as e:
-        logger.error(f"Error loading document: {str(e)}")
-        raise HTTPException(status_code=500, detail="Error loading document")
-
-    return document
